Add getAllAvailableIntervals and refactor Q1/2
getAllAvailableIntervals converts a list of WorkerAvailabilities into
a flat list of Intervals. Used this in a refactor of answers to Q1 and
Q2 to make them a little more sensible—now, rather than getting the
earliest/latest interval for each worker, then getting the
earliest/latest of those, we just get an array of *all* intervals, and
find the earliest/latest of that.